+++ src/App.jsx
@@ -7,7 +7,7 @@ import LandingPage from './Components/landingpage/HomePage';
 import LoginRegisterForm from './Components/LoginRegister/LoginRegister';
 import ViewReports from './Components/ViewReports';
 import Navbar from './Components/navbar/Navbar';
-import UserPage from './Components/profilePage/UserPage';
+import UserPage from './Components/UserPageComponent/UserPage';
 import Logout from './Components/Logout';
 import Dashboard from './Components/dashboard/dashboard';
 import Donation from './Components/donation/Donation';
